Pathing Testing
Seeing what would work better for moving the train through the environment.
Tried the physics constraint and found it a bit weird and difficult to work with.
Tried the spline and found that easier to work with overall.
Will probably do a spline follow for movement and add in a physics constraint only if the movement doesn't look good between carriages